refactor(workbench): drop duplicated demo pages and share the products demo builder
/tables/actions duplicated /products wholesale, /tables/behavior re-skinned
the home users table, /form/validation restated what ProductForm and
/form/dependent already demo, and /settings was an untested placeholder.
The product demo tables now share one base builder with the default sort.
