+++ tests/Browser/SmokeTest.php
@@ -19,15 +19,12 @@ it('renders every demo page without smoke failures', function (): void {
         '/form/fields/repeater',
         '/form/fields/builder',
         '/form/dependent',
-        '/form/validation',
         '/tables/columns/text',
         '/tables/columns/number',
         '/tables/columns/visual',
         '/tables/columns/custom',
         '/tables/filters',
         '/tables/pagination',
-        '/tables/actions',
-        '/tables/behavior',
         '/components/buttons',
         '/components/tabs',
         '/components/charts',
