+++ workbench/app/Tables/CustomColumnTable.php
@@ -7,16 +7,11 @@ use Illuminate\Database\Eloquent\Builder;
 use Lattice\Lattice\Attributes\AsTable;
 use Lattice\Lattice\Tables\Columns\Column;
 use Lattice\Lattice\Tables\Columns\TextColumn;
-use Lattice\Lattice\Tables\Sources\Eloquent\EloquentTableDefinition;
-use Lattice\Lattice\Tables\TableQuery;
 use Workbench\App\Models\Product;
 use Workbench\App\Tables\Columns\StatusBadgeColumn;
 
-/**
- * @extends EloquentTableDefinition<Product>
- */
 #[AsTable('workbench.demo.custom-column')]
-class CustomColumnTable extends EloquentTableDefinition
+class CustomColumnTable extends BaseProductsDemoTable
 {
     /**
      * @return array<int, Column>
@@ -36,14 +31,8 @@ class CustomColumnTable extends EloquentTableDefinition
     /**
      * @return Builder<Product>
      */
-    public function builder(TableQuery $query): Builder
+    protected function query(): Builder
     {
-        $builder = Product::query()->select(['id', 'name', 'status']);
-
-        if ($query->sorts === []) {
-            $builder->latest('id');
-        }
-
-        return $builder;
+        return Product::query()->select(['id', 'name', 'status']);
     }
 }
