+++ workbench/app/Tables/NumberColumnsTable.php
@@ -9,16 +9,11 @@ use Lattice\Lattice\Tables\Columns\Column;
 use Lattice\Lattice\Tables\Columns\MoneyColumn;
 use Lattice\Lattice\Tables\Columns\NumberColumn;
 use Lattice\Lattice\Tables\Columns\TextColumn;
-use Lattice\Lattice\Tables\Sources\Eloquent\EloquentTableDefinition;
-use Lattice\Lattice\Tables\TableQuery;
 use Workbench\App\Models\Product;
 use Workbench\App\Models\SalesPrice;
 
-/**
- * @extends EloquentTableDefinition<Product>
- */
 #[AsTable('workbench.demo.number-columns')]
-class NumberColumnsTable extends EloquentTableDefinition
+class NumberColumnsTable extends BaseProductsDemoTable
 {
     /**
      * @return array<int, Column>
@@ -35,9 +30,9 @@ class NumberColumnsTable extends EloquentTableDefinition
     /**
      * @return Builder<Product>
      */
-    public function builder(TableQuery $query): Builder
+    protected function query(): Builder
     {
-        $builder = Product::query()
+        return Product::query()
             ->select(['id', 'name'])
             ->selectSub(
                 SalesPrice::query()
@@ -53,11 +48,5 @@ class NumberColumnsTable extends EloquentTableDefinition
                     ->whereColumn('product_id', 'products.id'),
                 'prices_count',
             );
-
-        if ($query->sorts === []) {
-            $builder->latest('id');
-        }
-
-        return $builder;
     }
 }
